+++ js/main.js
@@ -4,3 +4,4 @@ import './posts-fullscreen.js'
 import './edit-image.js'
 import './comment-checking.js'
 import './api.js'
+import './popups.js'
